+++ src/wai/annotations/video/format/webcam/component/__init__.py
@@ -1,0 +1,1 @@
+from ._WebCamReader import WebCamReader