+++ src/wai/annotations/video/format/webcam/specifier/__init__.py
@@ -1,0 +1,1 @@
+from ._WebCamODInputFormatSpecifier import WebCamODInputFormatSpecifier