+++ CHANGES.rst
@@ -4,5 +4,5 @@ Changelog
 1.0.0 (????-??-??)
 ------------------
 
-- initial release: from-video-file-od
+- initial release: from-video-file-od, from-webcam-od